Set up connections between Dijkstra and Front-end
Current Issue:
There is an infinite loop somewhere in dijkstras.js (seemingly at getPath) that causes a RangeError when calling the function.

This is likely an issue with calling it, although the change that was made to dijkstras.js was changing it from logging in the console to a constantly updating string.